backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
sse-starlette==1.8.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis and job queue
redis==5.0.1
celery==5.3.4

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Image processing
Pillow==10.1.0

# Logging
loguru==0.7.2

# HTTP client
httpx==0.25.2

# Environment variables
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1